+++ character/views/mixins.py
@@ -1,15 +1,13 @@
-from django.core.cache import cache
 from django.http import Http404
 from django.views.generic import View
 from django.views.generic.list import ContextMixin
 
 from ..models.character import Character
-from ..utils.cache import character_key
 
 
 class CharacterContextMixin(ContextMixin, View):
     """
-    Mixin class that provides the character object and contextof the associated
+    Mixin class that provides the character object and context of the associated
     character's ID given in the view's URL.
 
     Attributes:
@@ -20,9 +18,7 @@ class CharacterContextMixin(ContextMixin, View):
         super().setup(request, *args, **kwargs)
         character_id = self.kwargs["character_id"]
         try:
-            self.character = cache.get_or_set(
-                character_key(character_id), Character.objects.get(id=character_id)
-            )
+            self.character = Character.objects.get(id=character_id)
         except Character.DoesNotExist as exc:
             raise Http404(f"Character of {character_id=} not found") from exc